create relation between book and author

+++ app/Http/Controllers/BookController.php
@@ -2,31 +2,32 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Author;
 use App\Models\Book;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class BookController extends Controller
 {
     public function index()
     {
-        return view('index');
+        $authors = Author::all();
+        return view('index', ['authors' => $authors]);
     }
 
     public function store(Request $request) {
-//        $bookData = [
-//            'title' => $request->title,
-//            'number' => $request->number,
-//            'description' => $request->description
-//        ];
         $bookData = $request->validate([
             'title' => ['required'],
-            'number' => ['required', Rule::unique('books', 'number')]
+            'number' => ['required', Rule::unique('books', 'number')],
+            'author_id' => ['required']
         ]);
         $bookData['description'] = $request->description;
+       $bookData['author_id'] = (int)$request->author_id;
+//       dd($bookData);
         Book::create($bookData);
         return response()->json([
             'status' => 200
@@ -45,6 +46,7 @@ class BookController extends Controller
                 <th>Title</th>
                 <th>Number</th>
                 <th>Description</th>
+                <th>Author</th>
                 <th>Action</th>
               </tr>
             </thead>
@@ -55,6 +57,7 @@ class BookController extends Controller
                 <td>' . $book->title . '</td>
                 <td>' . $book->number . '</td>
                 <td>' . $book->description . '</td>
+                <td>' . $book->author->name . '</td>
                 <td>
                   <a href="#" id="' . $book->id . '" class="text-success mx-1 editIcon" data-bs-toggle="modal" data-bs-target="#editBookModal"><i class="bi-pencil-square h4"></i></a>
 
@@ -69,7 +72,7 @@ class BookController extends Controller
         }
     }
 
-    //edit book
+    //edit book field inputs
     public function edit(Request $request) {
         $id = $request->id;
         $book = Book::find($id);
@@ -78,13 +81,13 @@ class BookController extends Controller
 
     // handle update an book and store data to database
     public function update(Request $request) {
-        $fileName = '';
         $book = Book::find($request->book_id);
 
         $bookData = [
             'title' => $request->title,
             'description' => $request->description,
-            'number' => $request->number
+            'number' => $request->number,
+            'author_id' => $request->author_id
         ];
         $book->update($bookData);
         return response()->json([
@@ -96,6 +99,9 @@ class BookController extends Controller
     public function delete(Request $request) {
         $id = $request->id;
         $book = Book::find($id);
+        if (!$book) {
+            throw new NotFoundHttpException('The book does not exist!');
+        }
         Book::destroy($id);
     }
 }
